requirements.txt
appnope==0.1.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
cycler==0.10.0
decorator==4.3.2
enum34==1.1.6
futures==3.2.0
ipykernel==4.10.0
ipython==5.8.0
ipython-genutils==0.2.0
jupyter-client==5.2.4
jupyter-core==4.4.0
kiwisolver==1.0.1
matplotlib==2.2.3
numpy==1.16.1
pandas==0.24.0
pathlib2==2.3.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.3.1
pyparsing==2.3.1
python-dateutil==2.7.5
pytz==2018.9
pyzmq==17.1.2
git+https://github.com/ctc316/roipoly.py.git
scandir==1.9.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
subprocess32==3.5.3
tornado==5.1.1
traitlets==4.3.2
wcwidth==0.1.7
